requirements.txt
opencv-python==4.7.0.72
numpy==1.26.3
pytesseract==0.3.10
pillow==10.2.0
